Adding the full code for figure 11
R script for figure 11 uploaded, along with the rds files to go with it.

Warning fully running the script by splitting the data and then re-fitting each model for each athlete will take a while to run!

I also changed the name of an rmd script which was incorrectly labelled.